feat: add control/setter api, refactor file contents, add setter api testings
